requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
python-magic>=0.4.27
